+++ vitest.shared.ts
@@ -2,7 +2,10 @@ import { defineConfig, type UserConfig } from 'vitest/config'
 
 export const packageTestInclude = ['__tests__/**/*.{spec,test}.ts']
 
-export const workspaceTestInclude = ['packages/*/__tests__/**/*.{spec,test}.ts']
+export const workspaceTestInclude = [
+  'packages/*/__tests__/**/*.{spec,test}.ts',
+  'scripts/**/__tests__/**/*.{spec,test}.ts',
+]
 
 export const workspaceCoverageInclude = ['packages/*/src/**/*.ts']